Fix HAWC periodic wind exhausted in IfW
There was a bug where the modulo operation to wrap back to the beginning
of the wind file didn't work correctly, previously noted in issue #471.
This seemed to come back due to some changes in the IfW refactor.
The ifw_HAWC regression test was added to catch this and other IfW
changes in the future.