HD/SeaSt: remove extra `WtrDens` and `RhoXg` copies
These are now stored in WaveField data since the modules that need them already access the rest of the WaveField data.
Also, all of the SeaSt InitOut data that are sent to HD Init are now done in one place instead of two (originally there was some issue with pointers in the SeaSt InitOut type, but that is not a concern any more).