[BugFix] FF mod_AmbWind=3 failed if a turbine not aligned exactly at Y0 of high res grid
If a turbine was shifted ever so slightly in the +/-Y direction from a high resolution grid point, the box of grid points sent to InflowWind for calculation would be shifted so one edge was just outside the wind box.  This would immediately fail.

For example, the FF input file:

```
--- WIND TURBINES ---
1             NumTurbines   - Number of wind turbines (-) [>=1] [last 6 columns below used only for Mod_AmbWind=2 or 3]
WT_X            WT_Y            WT_Z            WT_FASTInFile   X0_High         Y0_High         Z0_High         dX_High         dY_High         dZ_High
(m)             (m)             (m)             (string)        (m)             (m)             (m)             (m)             (m)             (m)
1.20200000e+02	-315.1	0	"FFTest_WT1.fst"	-1.00300000e+02	-540	1	4	5	5
```

shifts the turbine by 0.1 m in Y.  The Y0_high location is set to -540 m.  During calculations in AWAE, the call to InflowWind, the points were shifted by the turbine XY position.  It should have been shifted by the XY center of the box so that they aligned with the grid in the InflowWind calculations.

Updated the ModAmb_3 test case to verify this fixes the problem. Properly setting the offset to the center of the high resolution box also shifted the box 20 m in the X direction for the regression test, so the results changed very slightly.