Use more efficient method to compare linearization matrices in regression tests.
The previous method looped over each element and checked to see if it was within tolerance of the reference. This does the compare for the full matrices and then iterates over the element indices which were not within tolerance. This should significantly reduce the time to do the comparison for large matrices.

Also prints out the row/column variables which correspond to that element.