Fix MD_perturb_x indexing to get correct dx
The previous code passed in the state index to MD_perturb_x which then used that index to get dx from the array. However, the dx array is based on p%Jac_nx so the function was using the wrong dx to perturb the states. This change fixes it so the correct dx values are used for the appropriate states.