Debugging: is FileInfoType clobbering something else?
Observation -- read file ./tmp.dat using ProcessComFile inside the ReadBladeFile, and don't keep the data
 - at 64 lines, frequencies are ok and no huge acceleration values in OP
 - at 65 lines, something goes wrong and the frequencies change and huge accelerations in OP show up

Possibility 1:
Maybe the FileInfo%Lines is getting allocated to the stack even thought it should be on the heap at 65 lines?
This would result in clobbering something behind it on the stack that was already there?  Not sure that logic holds

Possibility 2:
At this size, it takes up enough space on the heap that when deallocated, something uninitialized ends up using that memory.  Problems with this include the "why has this never been observed before" question.