Increase number of unit numbers available for OMP file opening
We were having problems with the `!OMP` directives around high resolution file reading in AWAE.f90. Since the file reading starts with a call to `GetNewUnit` before starting the opening, parallel calls to `GetNewUnit` could result in the same unit number handed out to two processes.  The first process would open the file and start reading, but then the second process would open a different file with the same unit number causing read errors for both processes as they attempted to read the same file at the same time.

To resolve this, an optional start unit number is added to `GtNewUnit`.  This optional argument is propagated up through to the calls within the `!OMP` loop where a `N*100` is used as the starting unit number for the search (where `N` is the turbine number).  Theoretically this will solve the parallel file reading issue from AWAE.

Routine changes:

`GetNewUnit`
 - Increased maximum unit from 1024 to 64k (likely won't be an issue since most times where more than 1024 files are needed will be on supercomputers where the unit number limit is really high)
 - optional argument for starting unit number.  Defaults to old behaviour when not present.
`ReadLowResWindFile`
 - optional argument for starting unit number.  Defaults to old behaviour when not present.
`ReadHighResWindFile`
 - optional argument for starting unit number.  Defaults to old behaviour when not present.
AWAE `UpdateStates`
 - call to `ReadHighResWindFile` includes `nt*100` as the start unit number