FAST_Subs: fix missing code from PR #2267
Compare with https://github.com/OpenFAST/openfast/pull/2267

The summary file stated AeroDyn had unknown version and date instead of leaving that blank. I think the issue was that the original code on the AD14 removal had the `ELSEIF` indented improperly so the ELSE section was removed with the AD14 section.