AD bugfix: Segmentation fault with ifx
When OLAF was used with the AeroDyn Driver compiled with ifx (IFX) 2023.2.0 20230622 (release only), there would be a segmentation fault when SetInputsForFVW was called.  The root issue was that passing an array of `u` as `(/u/)` doesn't work correctly with this compiler.  So to work around this, the SetInputsForFVW routine was reworked to only operate on a single `u`.

In my opinion this is a hack of a solution to accomodate a compiler bug.