bugfix: IfW rotor points for disk average incorrect
The calculations for the rotor disk average wind speed use a set of point projected on the disk perpindular to the hub at ~0.7 rotor radius in the hub YZ plane.  However, the indices for this calculation were incorrect and projected the points onto disk in the hub XY plane.  This was discovered after one of the cases used in the curled wake paper failed to run correctly with 3.5.4 for a turbine very close to the back edge of the high resolution domain.

Also added some additional information about the location of points causing outside box bound errors.