MD: Allow external force and damping on MoorDyn body objects to be either defined in the global earth-fixed coordinate system or the local body-fixed coordinate system through a new input switch
External force and damping of point objects must always be in the global coordinate system due to a lack of rotational DoFs. The distributed transverse and axial damping of rod objects only make sense in the local body-fixed coordinate system. The behavior of external force on rod objects need to be clarified in the future.